+++ upup/pkg/fi/cloudup/do/api_target.go
@@ -17,17 +17,16 @@ limitations under the License.
 package do
 
 import (
-	"k8s.io/kops/pkg/resources/digitalocean"
 	"k8s.io/kops/upup/pkg/fi"
 )
 
 type DOAPITarget struct {
-	Cloud *digitalocean.Cloud
+	Cloud DOCloud
 }
 
 var _ fi.Target = &DOAPITarget{}
 
-func NewDOAPITarget(cloud *digitalocean.Cloud) *DOAPITarget {
+func NewDOAPITarget(cloud DOCloud) *DOAPITarget {
 	return &DOAPITarget{
 		Cloud: cloud,
 	}
