+++ upup/pkg/fi/cloudup/do/utils.go
@@ -1,0 +1,9 @@
+package do
+
+import "strings"
+
+func SafeClusterName(clusterName string) string {
+	// DO does not support . in tags / names
+	safeClusterName := strings.Replace(clusterName, ".", "-", -1)
+	return safeClusterName
+}
